ok but at least i figured out what i have to do next

+++ src/control/actions.rs
@@ -4,6 +4,6 @@ pub enum Action {
     Down,
     Left,
     Right,
-    Skill(u8),
+    Skill(usize),
     None
 }
